HHT-393: moving knowledge graph to separate package
