+++ app/rdf_transform/test_node_transformer.py
@@ -1,9 +1,9 @@
 from io import StringIO
 
-from app.rdf_transform.inmemory_knowledge_graph import InMemoryKnowledgeGraph
+from app.kg.inmemory_knowledge_graph import InMemoryKnowledgeGraph
+from app.kg.turtle_serializer import TurtleSerialializer
 from app.rdf_transform.node_transformer import NodesToRDFTransformer
 from app.rdf_transform.test_base import TransformBaseTest
-from app.rdf_transform.turtle_serializer import TurtleSerialializer
 
 
 class NodeTransformerTest(TransformBaseTest):
